requirements.txt
gradio
openai
Pillow
python-dotenv
pydantic
logging
base64
typing